fix: player hitbox

+++ src/entity/base/Entity.java
@@ -4,27 +4,27 @@ import sharedObject.Renderable;
 
 public abstract class Entity implements Renderable {
 
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 
-	public Entity(double x, double y) {
+	public Entity(int x, int y) {
 		this.setX(x);
 		this.setY(y);
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
